[proliferate] split session control actions
